[user-list-dropdown-plugin-implementation] - changes in review
