+++ bigbluebutton-html5/imports/ui/components/plugins-engine/plugin-provided-state/container.tsx
@@ -1,8 +1,9 @@
-import { useEffect, useState, useContext } from 'react';
+import * as React from 'react';
 import * as PluginSdk from 'bigbluebutton-html-plugin-sdk';
 
 import { PluginProvidedStateContainerProps, PluginsProvidedStateMap, PluginProvidedState } from '../types.ts';
-import { PluginsContext } from '../../components-data/plugin-context/context';
+import PresentationToolbarPluginStateContainer from './presentation-toolbar/container';
+import UserListDropdownPluginStateContainer from './user-list-dropdown/container';
 
 const pluginProvidedStateMap: PluginsProvidedStateMap = {};
 
@@ -21,56 +22,23 @@ const PluginProvidedStateContainer = (props: PluginProvidedStateContainerProps) 
     pluginProvidedStateMap[uuid] = {} as PluginProvidedState;
   }
   const pluginApi: PluginSdk.PluginApi = PluginSdk.getPluginApi(uuid);
-
-  const [
-    presentationToolbarItems,
-    setPresentationToolbarItems,
-  ] = useState<PluginSdk.PresentationToolbarItem[]>([]);
-
-  const [
-    userListDropdownItems,
-    setUserListDropdownItems,
-  ] = useState<PluginSdk.UserListDropdownItem[]>([]);
-
-  const {
-    pluginsProvidedAggregatedState,
-    setPluginsProvidedAggregatedState,
-  } = useContext(PluginsContext);
-
-  useEffect(() => {
-    // Change this plugin provided toolbar items
-    pluginProvidedStateMap[uuid].presentationToolbarItems = presentationToolbarItems;
-    pluginProvidedStateMap[uuid].userListDropdownItems = userListDropdownItems;
-
-    // Update context with computed aggregated list of all plugin provided toolbar items
-    const aggregatedPresentationToolbarItems = ([] as PluginSdk.PresentationToolbarItem[]).concat(
-      ...Object.values(pluginProvidedStateMap)
-        .map((pps: PluginProvidedState) => pps.presentationToolbarItems),
-    );
-    const aggregatedUserListDropdownItems = (
-      [] as PluginSdk.UserListDropdownItem[]).concat(
-      ...Object.values(pluginProvidedStateMap)
-        .map((pps: PluginProvidedState) => pps.userListDropdownItems),
-    );
-    setPluginsProvidedAggregatedState(
-      {
-        ...pluginsProvidedAggregatedState,
-        presentationToolbarItems: aggregatedPresentationToolbarItems,
-        userListDropdownItems: aggregatedUserListDropdownItems,
-      },
-    );
-  }, [presentationToolbarItems, userListDropdownItems]);
-
-  pluginApi.setPresentationToolbarItems = (items: PluginSdk.PresentationToolbarItem[]) => {
-    const itemsWithId = items.map(generateItemWithId);
-    return setPresentationToolbarItems(itemsWithId);
+  const pluginProvidedStateChildrenProps = {
+    uuid,
+    generateItemWithId,
+    pluginProvidedStateMap,
+    pluginApi,
   };
-
-  pluginApi.setUserListDropdownItems = (items: PluginSdk.UserListDropdownItem[]) => {
-    const itemsWithId = items.map(generateItemWithId);
-    return setUserListDropdownItems(itemsWithId);
-  };
-  return null;
+  console.log("Teste aqui ----> ", pluginProvidedStateChildrenProps)
+  return (
+    <>
+      <PresentationToolbarPluginStateContainer
+        { ...pluginProvidedStateChildrenProps}
+      />
+      <UserListDropdownPluginStateContainer
+        { ...pluginProvidedStateChildrenProps}
+      />
+    </>
+  );
 };
 
 export default PluginProvidedStateContainer;
